+++ szakdolgozat/szakdoga/objrepair.h
@@ -60,7 +60,7 @@ namespace approx{
 
 		//elemszam
 		int size() const {
-			return ind_map.size();
+			return (int)ind_map.size();
 		}
 
 		ConstIterator begin() const { return ConstIterator(&vecs,&ind_map,0); }
@@ -125,7 +125,7 @@ namespace approx{
 			inds.push_back(vmap[v]);
 		}
 		int transform_index(int ind) const { return inds[ind]; }
-		int size()const { return vecs.size(); }
+		int size()const { return (int)vecs.size(); }
 		Vector3<T> operator[](int ind) const {
 			return vecs[inds[ind]];
 		}
